refactor(icons): replace `Visibility` and `VisibilityOff` icons with grey SVG variants
- Updated `visibility-grey.svg` and `visibility-off-grey.svg` in `Resources`, `Field`, and related components.
- Removed unused `DashboardOverview` component.
